+++ lib/connect.js
@@ -441,6 +441,8 @@ function connect(token, options) {
  * @property {VideoRenderDimensions} [renderDimensions] - Optional parameter to specify the desired
  *   render dimensions of {@link RemoteVideoTrack}s based on {@link Track.Priority} and the
  *   {@link RemoteVideoTrack}s of the Dominant Speaker.
+ * @property {TrackSwitchOffMode} [trackSwitchOffMode="predicted"] - Optional parameter to configure
+ *   how {@link RemoteVideoTrack}s are switched off in response to bandwidth pressure. Defaults to "predicted".
  */
 
 /**
@@ -558,6 +560,31 @@ const NetworkQualityVerbosity = {
    * {@link Participant}'s {@link NetworkQualityLevel} is calculated.
    */
   detailed: 3
+};
+
+/**
+ * {@link TrackSwitchOffMode} specifies when {@link RemoteVideoTrack}s' are switched off.
+ * @enum {string}
+ */
+// eslint-disable-next-line
+const TrackSwitchOffMode = {
+  /**
+   * In this mode, {@link RemoteVideoTrack}s are switched off only when network congestion
+   * is detected.
+   */
+  detected: 'detected',
+
+  /**
+   * In this mode, {@link RemoteVideoTrack}s are pro-actively switched off when network
+   * congestion is predicted by the bandwidth estimation mechanism.
+   */
+  predicted: 'predicted',
+
+  /**
+   * In this mode, {@link RemoteVideoTrack}s are not switched off. Instead in response to network
+   * congestion, Tracks will be adjusted to lower quality.
+   */
+  disabled: 'disabled'
 };
 
 /**
