+++ lib/util/constants.js
@@ -86,6 +86,12 @@ module.exports.subscriptionMode = {
   MODE_PRESENTATION: 'presentation'
 };
 
+module.exports.trackSwitchOfMode = {
+  MODE_DISABLED: 'disabled',
+  MODE_DETECTED: 'detected',
+  MODE_PREDICTED: 'predicted'
+};
+
 module.exports.trackPriority = {
   PRIORITY_HIGH: 'high',
   PRIORITY_LOW: 'low',
