+++ lib/util/index.js
@@ -573,7 +573,8 @@ function createBandwidthProfileVideoPayload(bandwidthProfileVideo) {
     { prop: 'maxSubscriptionBitrate', type: 'number', payloadProp: 'max_subscription_bandwidth' },
     { prop: 'maxTracks', type: 'number', payloadProp: 'max_tracks' },
     { prop: 'mode', type: 'string' },
-    { prop: 'renderDimensions', type: 'object', payloadProp: 'render_dimensions', transform: createRenderDimensionsPayload }
+    { prop: 'renderDimensions', type: 'object', payloadProp: 'render_dimensions', transform: createRenderDimensionsPayload },
+    { prop: 'trackSwitchOffMode', type: 'string', payloadProp: 'track_switch_off' }
   ]);
 }
