+++ lib/util/validate.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { isNonArrayObject } = require('./');
-const { typeErrors: E, subscriptionMode, trackPriority } = require('./constants');
+const { typeErrors: E, subscriptionMode, trackPriority, trackSwitchOfMode } = require('./constants');
 
 /**
  * Validate the {@link BandwidthProfileOptions} object.
@@ -17,7 +17,8 @@ function validateBandwidthProfile(bandwidthProfile) {
     { prop: 'dominantSpeakerPriority', values: Object.values(trackPriority) },
     { prop: 'maxSubscriptionBitrate', type: 'number' },
     { prop: 'maxTracks', type: 'number' },
-    { prop: 'mode', values: Object.values(subscriptionMode) }
+    { prop: 'mode', values: Object.values(subscriptionMode) },
+    { prop: 'trackSwitchOffMode', values: Object.values(trackSwitchOfMode) }
   ]);
   return error || (bandwidthProfile.video
     ? validateRenderDimensions(bandwidthProfile.video.renderDimensions)
